+++ lib/service/DeviceCalendar.dart
@@ -1,15 +1,24 @@
 import 'package:device_calendar/device_calendar.dart';
+import 'package:illinois/model/Auth2.dart';
 import 'package:illinois/model/UserData.dart';
+import 'package:illinois/service/Auth2.dart';
+import 'package:illinois/service/NotificationService.dart';
 import 'package:illinois/service/Service.dart';
 import 'package:illinois/service/Storage.dart';
 import 'package:illinois/utils/Utils.dart';
 import 'package:illinois/model/Event.dart' as ExploreEvent;
 import 'package:url_launcher/url_launcher.dart';
 
-class DeviceCalendar with Service {
+class DeviceCalendar with Service implements NotificationsListener{
+
+  static const String notifyPromptPopupMessage    = "edu.illinois.rokwire.device_calendar.messaging.message.popup";
+  static const String notifyPlaceEventMessage     = "edu.illinois.rokwire.device_calendar.messaging.place.event";
+
   Calendar _defaultCalendar;
   Map<String, String> _calendarEventIdTable;
   DeviceCalendarPlugin _deviceCalendarPlugin;
+
+  bool _disableAdditionalDataLink = true;
 
   static final DeviceCalendar _instance = DeviceCalendar._internal();
 
@@ -30,26 +39,60 @@ class DeviceCalendar with Service {
     return false;
   }
 
-  Future<bool> addEvent(ExploreEvent.Event event) async{
-    String additionalUrl = _extractAdditionalDataUrl(event);
+  @override
+  void createService() {
+    NotificationService().subscribe(this, [
+      Auth2UserPrefs.notifyFavoriteChanged,
+      DeviceCalendar.notifyPlaceEventMessage
+    ]);
+  }
 
+  @override
+  void destroyService() {
+    NotificationService().unsubscribe(this);
+  }
+
+  Future<bool> addEvent(ExploreEvent.Event event) async{
     _debugToast("Add Event- iCall:${event.icalUrl}, outlook:${event.outlookUrl}, startDateLocal: ${event.startDateLocal}, endDateLocal: ${event.endDateLocal}");
+    
+    //User prefs
+    if(!canAddToCalendar){
+      _debugToast("Disabled");
+      return false;
+    }
+    
+    if(canShowPrompt){
+      _promptDialog(event);
+      return true;
+    }
+    
+    //URL
+    String additionalUrl = _disableAdditionalDataLink? null : _extractAdditionalDataUrl(event);
     if(AppString.isStringNotEmpty(additionalUrl)){
       _openAdditionalDataLink(additionalUrl);
       return true;
     }
 
+    return _placeCalendarEvent(event);
+  }
+
+  Future<bool> _placeCalendarEvent(ExploreEvent.Event event) async{
+
+    //PLUGIN
     if(_deviceCalendarPlugin == null){
-      await _initDeviceCalendarPlugin();
+      bool initResult = await _initDeviceCalendarPlugin();
+      if(!initResult ?? true){
+        _debugToast("Unable to init plugin");
+      }
     }
     _debugToast("Add to calendar- id:${_defaultCalendar.id}, name:${_defaultCalendar.name}, accountName:${_defaultCalendar.accountName}, accountType:${_defaultCalendar.accountType}, isReadOnly:${_defaultCalendar.isReadOnly}, isDefault:${_defaultCalendar.isDefault},");
-
+    //PERMISSIONS
     bool hasPermissions = await _requestPermissions();
 
     _debugToast("Has permissions: $hasPermissions");
+    //PLACE
     if(hasPermissions && _defaultCalendar!=null) {
       Event calendarEvent = _convertEvent(event);
-
       final createEventResult = await _deviceCalendarPlugin.createOrUpdateEvent(calendarEvent);
       if(createEventResult?.data!=null){
         _storeEventId(event.id, createEventResult?.data);
@@ -59,16 +102,20 @@ class DeviceCalendar with Service {
 
       if(!createEventResult.isSuccess) {
         AppToast.show(createEventResult?.data ?? createEventResult?.errorMessages ?? "Unable to save Event to calendar");
+        print(createEventResult?.errorMessages);
         return false;
       }
     }
-
+    
     return true;
   }
 
-  Future<bool> deleteEvent(event) async{
+  Future<bool> deleteEvent(ExploreEvent.Event event) async{
     if(_deviceCalendarPlugin == null){
-      await _initDeviceCalendarPlugin();
+      bool initResult = await _initDeviceCalendarPlugin();
+      if(!initResult ?? true){
+        _debugToast("Unable to init plugin");
+      }
     }
 
     String eventId = event?.id != null && _calendarEventIdTable!= null ? _calendarEventIdTable[event?.id] : null;
@@ -85,17 +132,17 @@ class DeviceCalendar with Service {
     return deleteEventResult?.isSuccess;
   }
 
-  Future<void> _initDeviceCalendarPlugin() async{
+  Future<bool> _initDeviceCalendarPlugin() async{
     _deviceCalendarPlugin = new DeviceCalendarPlugin();
     dynamic storedTable = Storage().calendarEventsTable ?? Map();
     _calendarEventIdTable = storedTable!=null ? Map<String, String>.from(storedTable): Map();
-    await _loadDefaultCalendar();
+    return await _loadDefaultCalendar();
   }
 
-  Future<void> _loadDefaultCalendar() async {
+  Future<bool> _loadDefaultCalendar() async {
     bool hasPermissions = await _requestPermissions();
     if(!hasPermissions) {
-      return;
+      return false;
     }
     final calendarsResult = await _deviceCalendarPlugin.retrieveCalendars();
     List<Calendar> calendars = calendarsResult.data;
@@ -103,8 +150,11 @@ class DeviceCalendar with Service {
       Calendar defaultCalendar = calendars.firstWhere((element) => element.isDefault);
       if (defaultCalendar!= null){
         _defaultCalendar = defaultCalendar;
+        return true;
       }
     }
+
+    return false;
   }
 
   Future<bool> _requestPermissions() async {
@@ -137,9 +187,8 @@ class DeviceCalendar with Service {
   }
 
   String _extractAdditionalDataUrl(ExploreEvent.Event event){
-//    return null; //TBD
     String additionalUrl = AppString.isStringNotEmpty(event.icalUrl) ? event.icalUrl : null;
-    additionalUrl = AppString.isStringNotEmpty(event.outlookUrl) ? event.outlookUrl : additionalUrl; //TBD decide do we support both at same time
+    additionalUrl = AppString.isStringNotEmpty(event.outlookUrl) ? event.outlookUrl : additionalUrl;
 
     return additionalUrl;
   }
@@ -159,5 +208,41 @@ class DeviceCalendar with Service {
 
   void _debugToast(String msg){
     AppToast.show(msg); //TBD Remove before release
+  }
+
+  void _processEvents(dynamic event){
+    if (event is ExploreEvent.Event) {
+      if (Auth2().isFavorite(event)) {
+        addEvent(event);
+      }
+      else {
+        deleteEvent(event);
+      }
+    }
+  }
+
+  void _promptDialog(ExploreEvent.Event event) {
+    NotificationService().notify(notifyPromptPopupMessage, event);
+  }
+
+
+
+  @override
+  void onNotification(String name, param) {
+    if(name == Auth2UserPrefs.notifyFavoriteChanged){
+      _processEvents(param);
+    } else if(name == DeviceCalendar.notifyPlaceEventMessage){
+      if(param!=null && param is ExploreEvent.Event){
+        _placeCalendarEvent(param);
+      }
+    }
+  }
+  
+  bool get canAddToCalendar{
+    return Storage().calendarEnabledToSave ?? false;
+  }
+  
+  bool get canShowPrompt{
+    return Storage().calendarCanPrompt ?? false;
   }
 }